requirements.txt
python-gitlab

